refactor: deepen scaffold dialects, layouts, and codegen seams
Collapse duplicated validator/framework scaffolding into dialect-owned
field formatting, layout assemblers, and a shared Next/Nuxt factory so
prefix/package config and template structure live in one place each.
