+++ packages/arkenv/src/features/scaffold/index.ts
@@ -1,5 +1,5 @@
-export * from "./env-template";
 export * from "./executor";
 export * from "./plan";
 export * from "./planner";
 export * from "./scaffold";
+export type { StrictEnvTemplates } from "./validators";
