+++ packages/arkenv/src/features/scaffold/validators/index.ts
@@ -1,7 +1,9 @@
-import { arktypeStrategy } from "./arktype";
+import {
+	arktypeStrategy,
+	valibotStrategy,
+	zodStrategy,
+} from "./create-strategy";
 import type { ValidatorRegistry } from "./types";
-import { valibotStrategy } from "./valibot";
-import { zodStrategy } from "./zod";
 
 /**
  * Exhaustive registry of validator strategies.
@@ -13,3 +15,5 @@ export const VALIDATORS = {
 } satisfies ValidatorRegistry;
 
 export type { StrictEnvTemplates, ValidatorStrategy } from "./types";
+export type { Dialect } from "./dialects";
+export { DIALECTS } from "./dialects";
